+++ ESafety.Core.Model/PARA/Dict.cs
@@ -38,11 +38,11 @@ namespace ESafety.Core.Model.PARA
         /// <summary>
         /// 最小值
         /// </summary>
-        public int MinValue { get; set; }
+        public double MinValue { get; set; }
         /// <summary>
         /// 最大值
         /// </summary>
-        public int MaxValue { get; set; }
+        public double MaxValue { get; set; }
     }
     /// <summary>
     /// 修改词典
